encrypt not working

+++ lib/lab10/inc/RSAE.h
@@ -24,8 +24,8 @@ namespace lab10 {
         void generate_keys();
         //todo: encrypt message
         //whole number numerical ecryption/decryption
-        void encrypt(unsigned message, std::string key);
-        void decrypt(double  message, std::string key);
+        void encrypt(long double message, std::string key);
+        void decrypt(long double  message, std::string key);
         //todo: decrypt message
         //I wanna do this shit strings,
         //if can, with text files
